+++ clang-tools-extra/clang-tidy/bugprone/StringviewNullptrCheck.h
@@ -9,7 +9,7 @@
 #ifndef LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_BUGPRONE_STRINGVIEWNULLPTRCHECK_H
 #define LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_BUGPRONE_STRINGVIEWNULLPTRCHECK_H
 
-#include "../utils/TransformerClangTidyCheck.h"
+#include "../ClangTidyCheck.h"
 
 namespace clang::tidy::bugprone {
 
@@ -19,21 +19,13 @@ namespace clang::tidy::bugprone {
 /// braced initializer list does not compile so instead a call to `.empty()` or
 /// the empty string literal are used, where appropriate.
 ///
-/// This prevents code from invoking behavior which is unconditionally
-/// undefined. The single-argument `const CharT*` constructor does not check
-/// for the null case before dereferencing its input. The standard is slated to
-/// add an explicitly-deleted overload to catch some of these cases:
-/// wg21.link/p2166
-///
-/// To catch the additional cases of `NULL` (which expands to `__null`) and
-/// `0`, first run the ``modernize-use-nullptr`` check to convert the callers
-/// to `nullptr`.
-///
 /// For the user-facing documentation see:
 /// https://clang.llvm.org/extra/clang-tidy/checks/bugprone/stringview-nullptr.html
-class StringviewNullptrCheck : public utils::TransformerClangTidyCheck {
+class StringviewNullptrCheck : public ClangTidyCheck {
 public:
-  StringviewNullptrCheck(StringRef Name, ClangTidyContext *Context);
+  using ClangTidyCheck::ClangTidyCheck;
+  void registerMatchers(ast_matchers::MatchFinder *Finder) override;
+  void check(const ast_matchers::MatchFinder::MatchResult &Result) override;
 
   bool isLanguageVersionSupported(const LangOptions &LangOpts) const override {
     return LangOpts.CPlusPlus17;